Refine set password
