+++ web-v4/src/routes/settings/+page.server.ts
@@ -1,5 +1,6 @@
 import { env } from '$env/dynamic/private';
 import { requestEmailChange, deleteAccount } from '$lib/data/settings';
+import { setPassword } from '$lib/data/session';
 import { getCurrentUser } from '$lib/data/user';
 import { getHTTPErrors } from '$lib/utils/actions';
 import { safeAsync } from '$lib/utils/async';
@@ -27,7 +28,32 @@ export const actions = {
 
 		try {
 			await requestEmailChange(email.trim(), host, locals);
-			return { success: true, email: email.trim() };
+			return { emailSuccess: true, email: email.trim() };
+		} catch (error) {
+			return fail(422, getHTTPErrors(error));
+		}
+	},
+
+	changePassword: async ({ request, locals }) => {
+		const formData = await request.formData();
+		const data = Object.fromEntries(formData);
+
+		const password = data.password as string;
+		if (!password || password.length < 8) {
+			return fail(422, {
+				errors: [{ id: 'password', title: 'Password must be at least 8 characters' }]
+			});
+		}
+
+		if (password !== data.confirm_password) {
+			return fail(422, {
+				errors: [{ id: 'confirm_password', title: 'Passwords do not match' }]
+			});
+		}
+
+		try {
+			await setPassword({ password }, locals);
+			return { passwordSuccess: true };
 		} catch (error) {
 			return fail(422, getHTTPErrors(error));
 		}
